Refresh with moved GapicClient objects
