+++ src/ErrorReporting/V1beta1/ErrorGroupServiceClient.php
@@ -30,6 +30,8 @@
 
 namespace Google\Cloud\ErrorReporting\V1beta1;
 
+use Google\Cloud\ErrorReporting\V1beta1\Gapic\ErrorGroupServiceGapicClient;
+
 /**
  * {@inheritdoc}
  */
