+++ src/Logging/V2/Gapic/ConfigServiceV2GapicClient.php
@@ -18,7 +18,7 @@
 /*
  * GENERATED CODE WARNING
  * This file was generated from the file
- * https://github.com/google/googleapis/blob/master/google/logging/v2/logging_metrics.proto
+ * https://github.com/google/googleapis/blob/master/google/logging/v2/logging_config.proto
  * and updates to that file get reflected here through a refresh process.
  *
  * EXPERIMENTAL: this client library class has not yet been declared beta. This class may change
@@ -28,7 +28,7 @@
  * @experimental
  */
 
-namespace Google\Cloud\Logging\V2;
+namespace Google\Cloud\Logging\V2\Gapic;
 
 use Google\GAX\AgentHeaderDescriptor;
 use Google\GAX\ApiCallable;
@@ -37,16 +37,17 @@ use Google\GAX\GrpcConstants;
 use Google\GAX\GrpcCredentialsHelper;
 use Google\GAX\PageStreamingDescriptor;
 use Google\GAX\PathTemplate;
-use Google\Logging\V2\CreateLogMetricRequest;
-use Google\Logging\V2\DeleteLogMetricRequest;
-use Google\Logging\V2\GetLogMetricRequest;
-use Google\Logging\V2\ListLogMetricsRequest;
-use Google\Logging\V2\LogMetric;
-use Google\Logging\V2\MetricsServiceV2GrpcClient;
-use Google\Logging\V2\UpdateLogMetricRequest;
+use Google\Logging\V2\ConfigServiceV2GrpcClient;
+use Google\Logging\V2\CreateSinkRequest;
+use Google\Logging\V2\DeleteSinkRequest;
+use Google\Logging\V2\GetSinkRequest;
+use Google\Logging\V2\ListSinksRequest;
+use Google\Logging\V2\LogSink;
+use Google\Logging\V2\UpdateSinkRequest;
 
 /**
- * Service Description: Service for configuring logs-based metrics.
+ * Service Description: Service for configuring sinks used to export log entries outside of
+ * Stackdriver Logging.
  *
  * EXPERIMENTAL: this client library class has not yet been declared beta. This class may change
  * more frequently than those which have been declared beta or 1.0, including changes which break
@@ -57,23 +58,23 @@ use Google\Logging\V2\UpdateLogMetricRequest;
  *
  * ```
  * try {
- *     $metricsServiceV2Client = new MetricsServiceV2Client();
- *     $formattedParent = MetricsServiceV2Client::formatProjectName("[PROJECT]");
+ *     $configServiceV2Client = new ConfigServiceV2Client();
+ *     $formattedParent = ConfigServiceV2Client::formatProjectName("[PROJECT]");
  *     // Iterate through all elements
- *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent);
+ *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent);
  *     foreach ($pagedResponse->iterateAllElements() as $element) {
  *         // doSomethingWith($element);
  *     }
  *
  *     // OR iterate over pages of elements, with the maximum page size set to 5
- *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent, ['pageSize' => 5]);
+ *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent, ['pageSize' => 5]);
  *     foreach ($pagedResponse->iteratePages() as $page) {
  *         foreach ($page as $element) {
  *             // doSomethingWith($element);
  *         }
  *     }
  * } finally {
- *     $metricsServiceV2Client->close();
+ *     $configServiceV2Client->close();
  * }
  * ```
  *
@@ -84,7 +85,7 @@ use Google\Logging\V2\UpdateLogMetricRequest;
  *
  * @experimental
  */
-class MetricsServiceV2GapicClient
+class ConfigServiceV2GapicClient
 {
     /**
      * The default address of the service.
@@ -112,10 +113,10 @@ class MetricsServiceV2GapicClient
     const CODEGEN_VERSION = '0.0.5';
 
     private static $projectNameTemplate;
-    private static $metricNameTemplate;
+    private static $sinkNameTemplate;
 
     protected $grpcCredentialsHelper;
-    protected $metricsServiceV2Stub;
+    protected $configServiceV2Stub;
     private $scopes;
     private $defaultCallSettings;
     private $descriptors;
@@ -138,19 +139,19 @@ class MetricsServiceV2GapicClient
 
     /**
      * Formats a string containing the fully-qualified path to represent
-     * a metric resource.
+     * a sink resource.
      *
      * @param string $project
-     * @param string $metric
+     * @param string $sink
      *
-     * @return string The formatted metric resource.
+     * @return string The formatted sink resource.
      * @experimental
      */
-    public static function formatMetricName($project, $metric)
+    public static function formatSinkName($project, $sink)
     {
-        return self::getMetricNameTemplate()->render([
+        return self::getSinkNameTemplate()->render([
             'project' => $project,
-            'metric' => $metric,
+            'sink' => $sink,
         ]);
     }
 
@@ -170,30 +171,30 @@ class MetricsServiceV2GapicClient
 
     /**
      * Parses the project from the given fully-qualified path which
-     * represents a metric resource.
+     * represents a sink resource.
      *
-     * @param string $metricName The fully-qualified metric resource.
+     * @param string $sinkName The fully-qualified sink resource.
      *
      * @return string The extracted project value.
      * @experimental
      */
-    public static function parseProjectFromMetricName($metricName)
+    public static function parseProjectFromSinkName($sinkName)
     {
-        return self::getMetricNameTemplate()->match($metricName)['project'];
+        return self::getSinkNameTemplate()->match($sinkName)['project'];
     }
 
     /**
-     * Parses the metric from the given fully-qualified path which
-     * represents a metric resource.
+     * Parses the sink from the given fully-qualified path which
+     * represents a sink resource.
      *
-     * @param string $metricName The fully-qualified metric resource.
+     * @param string $sinkName The fully-qualified sink resource.
      *
-     * @return string The extracted metric value.
+     * @return string The extracted sink value.
      * @experimental
      */
-    public static function parseMetricFromMetricName($metricName)
+    public static function parseSinkFromSinkName($sinkName)
     {
-        return self::getMetricNameTemplate()->match($metricName)['metric'];
+        return self::getSinkNameTemplate()->match($sinkName)['sink'];
     }
 
     private static function getProjectNameTemplate()
@@ -205,29 +206,29 @@ class MetricsServiceV2GapicClient
         return self::$projectNameTemplate;
     }
 
-    private static function getMetricNameTemplate()
+    private static function getSinkNameTemplate()
     {
-        if (self::$metricNameTemplate == null) {
-            self::$metricNameTemplate = new PathTemplate('projects/{project}/metrics/{metric}');
+        if (self::$sinkNameTemplate == null) {
+            self::$sinkNameTemplate = new PathTemplate('projects/{project}/sinks/{sink}');
         }
 
-        return self::$metricNameTemplate;
+        return self::$sinkNameTemplate;
     }
 
     private static function getPageStreamingDescriptors()
     {
-        $listLogMetricsPageStreamingDescriptor =
+        $listSinksPageStreamingDescriptor =
                 new PageStreamingDescriptor([
                     'requestPageTokenGetMethod' => 'getPageToken',
                     'requestPageTokenSetMethod' => 'setPageToken',
                     'requestPageSizeGetMethod' => 'getPageSize',
                     'requestPageSizeSetMethod' => 'setPageSize',
                     'responsePageTokenGetMethod' => 'getNextPageToken',
-                    'resourcesGetMethod' => 'getMetrics',
+                    'resourcesGetMethod' => 'getSinks',
                 ]);
 
         $pageStreamingDescriptors = [
-            'listLogMetrics' => $listLogMetricsPageStreamingDescriptor,
+            'listSinks' => $listSinksPageStreamingDescriptor,
         ];
 
         return $pageStreamingDescriptors;
@@ -309,22 +310,22 @@ class MetricsServiceV2GapicClient
 
         $defaultDescriptors = ['headerDescriptor' => $headerDescriptor];
         $this->descriptors = [
-            'listLogMetrics' => $defaultDescriptors,
-            'getLogMetric' => $defaultDescriptors,
-            'createLogMetric' => $defaultDescriptors,
-            'updateLogMetric' => $defaultDescriptors,
-            'deleteLogMetric' => $defaultDescriptors,
+            'listSinks' => $defaultDescriptors,
+            'getSink' => $defaultDescriptors,
+            'createSink' => $defaultDescriptors,
+            'updateSink' => $defaultDescriptors,
+            'deleteSink' => $defaultDescriptors,
         ];
         $pageStreamingDescriptors = self::getPageStreamingDescriptors();
         foreach ($pageStreamingDescriptors as $method => $pageStreamingDescriptor) {
             $this->descriptors[$method]['pageStreamingDescriptor'] = $pageStreamingDescriptor;
         }
 
-        $clientConfigJsonString = file_get_contents(__DIR__.'/resources/metrics_service_v2_client_config.json');
+        $clientConfigJsonString = file_get_contents(__DIR__.'/../resources/config_service_v2_client_config.json');
         $clientConfig = json_decode($clientConfigJsonString, true);
         $this->defaultCallSettings =
                 CallSettings::load(
-                    'google.logging.v2.MetricsServiceV2',
+                    'google.logging.v2.ConfigServiceV2',
                     $clientConfig,
                     $options['retryingOverride'],
                     GrpcConstants::getStatusCodeNames(),
@@ -339,44 +340,47 @@ class MetricsServiceV2GapicClient
         }
         $this->grpcCredentialsHelper = new GrpcCredentialsHelper($options);
 
-        $createMetricsServiceV2StubFunction = function ($hostname, $opts, $channel) {
-            return new MetricsServiceV2GrpcClient($hostname, $opts, $channel);
+        $createConfigServiceV2StubFunction = function ($hostname, $opts, $channel) {
+            return new ConfigServiceV2GrpcClient($hostname, $opts, $channel);
         };
-        if (array_key_exists('createMetricsServiceV2StubFunction', $options)) {
-            $createMetricsServiceV2StubFunction = $options['createMetricsServiceV2StubFunction'];
+        if (array_key_exists('createConfigServiceV2StubFunction', $options)) {
+            $createConfigServiceV2StubFunction = $options['createConfigServiceV2StubFunction'];
         }
-        $this->metricsServiceV2Stub = $this->grpcCredentialsHelper->createStub($createMetricsServiceV2StubFunction);
+        $this->configServiceV2Stub = $this->grpcCredentialsHelper->createStub($createConfigServiceV2StubFunction);
     }
 
     /**
-     * Lists logs-based metrics.
+     * Lists sinks.
      *
      * Sample code:
      * ```
      * try {
-     *     $metricsServiceV2Client = new MetricsServiceV2Client();
-     *     $formattedParent = MetricsServiceV2Client::formatProjectName("[PROJECT]");
+     *     $configServiceV2Client = new ConfigServiceV2Client();
+     *     $formattedParent = ConfigServiceV2Client::formatProjectName("[PROJECT]");
      *     // Iterate through all elements
-     *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent);
+     *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent);
      *     foreach ($pagedResponse->iterateAllElements() as $element) {
      *         // doSomethingWith($element);
      *     }
      *
      *     // OR iterate over pages of elements, with the maximum page size set to 5
-     *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent, ['pageSize' => 5]);
+     *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent, ['pageSize' => 5]);
      *     foreach ($pagedResponse->iteratePages() as $page) {
      *         foreach ($page as $element) {
      *             // doSomethingWith($element);
      *         }
      *     }
      * } finally {
-     *     $metricsServiceV2Client->close();
+     *     $configServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $parent Required. The name of the project containing the metrics:
+     * @param string $parent Required. The parent resource whose sinks are to be listed:
      *
      *     "projects/[PROJECT_ID]"
+     *     "organizations/[ORGANIZATION_ID]"
+     *     "billingAccounts/[BILLING_ACCOUNT_ID]"
+     *     "folders/[FOLDER_ID]"
      * @param array $optionalArgs {
      *                            Optional.
      *
@@ -402,9 +406,9 @@ class MetricsServiceV2GapicClient
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function listLogMetrics($parent, $optionalArgs = [])
+    public function listSinks($parent, $optionalArgs = [])
     {
-        $request = new ListLogMetricsRequest();
+        $request = new ListSinksRequest();
         $request->setParent($parent);
         if (isset($optionalArgs['pageToken'])) {
             $request->setPageToken($optionalArgs['pageToken']);
@@ -413,14 +417,14 @@ class MetricsServiceV2GapicClient
             $request->setPageSize($optionalArgs['pageSize']);
         }
 
-        $mergedSettings = $this->defaultCallSettings['listLogMetrics']->merge(
+        $mergedSettings = $this->defaultCallSettings['listSinks']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->metricsServiceV2Stub,
-            'ListLogMetrics',
+            $this->configServiceV2Stub,
+            'ListSinks',
             $mergedSettings,
-            $this->descriptors['listLogMetrics']
+            $this->descriptors['listSinks']
         );
 
         return $callable(
@@ -430,22 +434,27 @@ class MetricsServiceV2GapicClient
     }
 
     /**
-     * Gets a logs-based metric.
+     * Gets a sink.
      *
      * Sample code:
      * ```
      * try {
-     *     $metricsServiceV2Client = new MetricsServiceV2Client();
-     *     $formattedMetricName = MetricsServiceV2Client::formatMetricName("[PROJECT]", "[METRIC]");
-     *     $response = $metricsServiceV2Client->getLogMetric($formattedMetricName);
+     *     $configServiceV2Client = new ConfigServiceV2Client();
+     *     $formattedSinkName = ConfigServiceV2Client::formatSinkName("[PROJECT]", "[SINK]");
+     *     $response = $configServiceV2Client->getSink($formattedSinkName);
      * } finally {
-     *     $metricsServiceV2Client->close();
+     *     $configServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $metricName The resource name of the desired metric:
+     * @param string $sinkName Required. The resource name of the sink:
      *
-     *     "projects/[PROJECT_ID]/metrics/[METRIC_ID]"
+     *     "projects/[PROJECT_ID]/sinks/[SINK_ID]"
+     *     "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]"
+     *     "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]"
+     *     "folders/[FOLDER_ID]/sinks/[SINK_ID]"
+     *
+     * Example: `"projects/my-project-id/sinks/my-sink-id"`.
      * @param array $optionalArgs {
      *                            Optional.
      *
@@ -457,24 +466,24 @@ class MetricsServiceV2GapicClient
      *          is not set.
      * }
      *
-     * @return \Google\Logging\V2\LogMetric
+     * @return \Google\Logging\V2\LogSink
      *
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function getLogMetric($metricName, $optionalArgs = [])
+    public function getSink($sinkName, $optionalArgs = [])
     {
-        $request = new GetLogMetricRequest();
-        $request->setMetricName($metricName);
+        $request = new GetSinkRequest();
+        $request->setSinkName($sinkName);
 
-        $mergedSettings = $this->defaultCallSettings['getLogMetric']->merge(
+        $mergedSettings = $this->defaultCallSettings['getSink']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->metricsServiceV2Stub,
-            'GetLogMetric',
+            $this->configServiceV2Stub,
+            'GetSink',
             $mergedSettings,
-            $this->descriptors['getLogMetric']
+            $this->descriptors['getSink']
         );
 
         return $callable(
@@ -484,30 +493,49 @@ class MetricsServiceV2GapicClient
     }
 
     /**
-     * Creates a logs-based metric.
+     * Creates a sink that exports specified log entries to a destination.  The
+     * export of newly-ingested log entries begins immediately, unless the current
+     * time is outside the sink's start and end times or the sink's
+     * `writer_identity` is not permitted to write to the destination.  A sink can
+     * export log entries only from the resource owning the sink.
      *
      * Sample code:
      * ```
      * try {
-     *     $metricsServiceV2Client = new MetricsServiceV2Client();
-     *     $formattedParent = MetricsServiceV2Client::formatProjectName("[PROJECT]");
-     *     $metric = new LogMetric();
-     *     $response = $metricsServiceV2Client->createLogMetric($formattedParent, $metric);
+     *     $configServiceV2Client = new ConfigServiceV2Client();
+     *     $formattedParent = ConfigServiceV2Client::formatProjectName("[PROJECT]");
+     *     $sink = new LogSink();
+     *     $response = $configServiceV2Client->createSink($formattedParent, $sink);
      * } finally {
-     *     $metricsServiceV2Client->close();
+     *     $configServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $parent The resource name of the project in which to create the metric:
+     * @param string $parent Required. The resource in which to create the sink:
      *
      *     "projects/[PROJECT_ID]"
+     *     "organizations/[ORGANIZATION_ID]"
+     *     "billingAccounts/[BILLING_ACCOUNT_ID]"
+     *     "folders/[FOLDER_ID]"
      *
-     * The new metric must be provided in the request.
-     * @param LogMetric $metric       The new logs-based metric, which must not have an identifier that
-     *                                already exists.
-     * @param array     $optionalArgs {
-     *                                Optional.
+     * Examples: `"projects/my-logging-project"`, `"organizations/123456789"`.
+     * @param LogSink $sink         Required. The new sink, whose `name` parameter is a sink identifier that
+     *                              is not already in use.
+     * @param array   $optionalArgs {
+     *                              Optional.
      *
+     *     @type bool $uniqueWriterIdentity
+     *          Optional. Determines the kind of IAM identity returned as `writer_identity`
+     *          in the new sink.  If this value is omitted or set to false, and if the
+     *          sink's parent is a project, then the value returned as `writer_identity` is
+     *          the same group or service account used by Stackdriver Logging before the
+     *          addition of writer identities to this API. The sink's destination must be
+     *          in the same project as the sink itself.
+     *
+     *          If this field is set to true, or if the sink is owned by a non-project
+     *          resource such as an organization, then the value of `writer_identity` will
+     *          be a unique service account used only for exports from the new sink.  For
+     *          more information, see `writer_identity` in [LogSink][google.logging.v2.LogSink].
      *     @type \Google\GAX\RetrySettings $retrySettings
      *          Retry settings to use for this call. If present, then
      *          $timeoutMillis is ignored.
@@ -516,25 +544,28 @@ class MetricsServiceV2GapicClient
      *          is not set.
      * }
      *
-     * @return \Google\Logging\V2\LogMetric
+     * @return \Google\Logging\V2\LogSink
      *
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function createLogMetric($parent, $metric, $optionalArgs = [])
+    public function createSink($parent, $sink, $optionalArgs = [])
     {
-        $request = new CreateLogMetricRequest();
+        $request = new CreateSinkRequest();
         $request->setParent($parent);
-        $request->setMetric($metric);
+        $request->setSink($sink);
+        if (isset($optionalArgs['uniqueWriterIdentity'])) {
+            $request->setUniqueWriterIdentity($optionalArgs['uniqueWriterIdentity']);
+        }
 
-        $mergedSettings = $this->defaultCallSettings['createLogMetric']->merge(
+        $mergedSettings = $this->defaultCallSettings['createSink']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->metricsServiceV2Stub,
-            'CreateLogMetric',
+            $this->configServiceV2Stub,
+            'CreateSink',
             $mergedSettings,
-            $this->descriptors['createLogMetric']
+            $this->descriptors['createSink']
         );
 
         return $callable(
@@ -544,31 +575,54 @@ class MetricsServiceV2GapicClient
     }
 
     /**
-     * Creates or updates a logs-based metric.
+     * Updates a sink. If the named sink doesn't exist, then this method is
+     * identical to
+     * [sinks.create](https://cloud.google.com/logging/docs/api/reference/rest/v2/projects.sinks/create).
+     * If the named sink does exist, then this method replaces the following
+     * fields in the existing sink with values from the new sink: `destination`,
+     * `filter`, `output_version_format`, `start_time`, and `end_time`.
+     * The updated filter might also have a new `writer_identity`; see the
+     * `unique_writer_identity` field.
      *
      * Sample code:
      * ```
      * try {
-     *     $metricsServiceV2Client = new MetricsServiceV2Client();
-     *     $formattedMetricName = MetricsServiceV2Client::formatMetricName("[PROJECT]", "[METRIC]");
-     *     $metric = new LogMetric();
-     *     $response = $metricsServiceV2Client->updateLogMetric($formattedMetricName, $metric);
+     *     $configServiceV2Client = new ConfigServiceV2Client();
+     *     $formattedSinkName = ConfigServiceV2Client::formatSinkName("[PROJECT]", "[SINK]");
+     *     $sink = new LogSink();
+     *     $response = $configServiceV2Client->updateSink($formattedSinkName, $sink);
      * } finally {
-     *     $metricsServiceV2Client->close();
+     *     $configServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $metricName The resource name of the metric to update:
+     * @param string $sinkName Required. The full resource name of the sink to update, including the
+     *                         parent resource and the sink identifier:
      *
-     *     "projects/[PROJECT_ID]/metrics/[METRIC_ID]"
+     *     "projects/[PROJECT_ID]/sinks/[SINK_ID]"
+     *     "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]"
+     *     "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]"
+     *     "folders/[FOLDER_ID]/sinks/[SINK_ID]"
      *
-     * The updated metric must be provided in the request and it's
-     * `name` field must be the same as `[METRIC_ID]` If the metric
-     * does not exist in `[PROJECT_ID]`, then a new metric is created.
-     * @param LogMetric $metric       The updated metric.
-     * @param array     $optionalArgs {
-     *                                Optional.
+     * Example: `"projects/my-project-id/sinks/my-sink-id"`.
+     * @param LogSink $sink         Required. The updated sink, whose name is the same identifier that appears
+     *                              as part of `sink_name`.  If `sink_name` does not exist, then
+     *                              this method creates a new sink.
+     * @param array   $optionalArgs {
+     *                              Optional.
      *
+     *     @type bool $uniqueWriterIdentity
+     *          Optional. See
+     *          [sinks.create](https://cloud.google.com/logging/docs/api/reference/rest/v2/projects.sinks/create)
+     *          for a description of this field.  When updating a sink, the effect of this
+     *          field on the value of `writer_identity` in the updated sink depends on both
+     *          the old and new values of this field:
+     *
+     *          +   If the old and new values of this field are both false or both true,
+     *              then there is no change to the sink's `writer_identity`.
+     *          +   If the old value is false and the new value is true, then
+     *              `writer_identity` is changed to a unique service account.
+     *          +   It is an error if the old value is true and the new value is false.
      *     @type \Google\GAX\RetrySettings $retrySettings
      *          Retry settings to use for this call. If present, then
      *          $timeoutMillis is ignored.
@@ -577,25 +631,28 @@ class MetricsServiceV2GapicClient
      *          is not set.
      * }
      *
-     * @return \Google\Logging\V2\LogMetric
+     * @return \Google\Logging\V2\LogSink
      *
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function updateLogMetric($metricName, $metric, $optionalArgs = [])
+    public function updateSink($sinkName, $sink, $optionalArgs = [])
     {
-        $request = new UpdateLogMetricRequest();
-        $request->setMetricName($metricName);
-        $request->setMetric($metric);
+        $request = new UpdateSinkRequest();
+        $request->setSinkName($sinkName);
+        $request->setSink($sink);
+        if (isset($optionalArgs['uniqueWriterIdentity'])) {
+            $request->setUniqueWriterIdentity($optionalArgs['uniqueWriterIdentity']);
+        }
 
-        $mergedSettings = $this->defaultCallSettings['updateLogMetric']->merge(
+        $mergedSettings = $this->defaultCallSettings['updateSink']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->metricsServiceV2Stub,
-            'UpdateLogMetric',
+            $this->configServiceV2Stub,
+            'UpdateSink',
             $mergedSettings,
-            $this->descriptors['updateLogMetric']
+            $this->descriptors['updateSink']
         );
 
         return $callable(
@@ -605,22 +662,29 @@ class MetricsServiceV2GapicClient
     }
 
     /**
-     * Deletes a logs-based metric.
+     * Deletes a sink. If the sink has a unique `writer_identity`, then that
+     * service account is also deleted.
      *
      * Sample code:
      * ```
      * try {
-     *     $metricsServiceV2Client = new MetricsServiceV2Client();
-     *     $formattedMetricName = MetricsServiceV2Client::formatMetricName("[PROJECT]", "[METRIC]");
-     *     $metricsServiceV2Client->deleteLogMetric($formattedMetricName);
+     *     $configServiceV2Client = new ConfigServiceV2Client();
+     *     $formattedSinkName = ConfigServiceV2Client::formatSinkName("[PROJECT]", "[SINK]");
+     *     $configServiceV2Client->deleteSink($formattedSinkName);
      * } finally {
-     *     $metricsServiceV2Client->close();
+     *     $configServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $metricName The resource name of the metric to delete:
+     * @param string $sinkName Required. The full resource name of the sink to delete, including the
+     *                         parent resource and the sink identifier:
      *
-     *     "projects/[PROJECT_ID]/metrics/[METRIC_ID]"
+     *     "projects/[PROJECT_ID]/sinks/[SINK_ID]"
+     *     "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]"
+     *     "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]"
+     *     "folders/[FOLDER_ID]/sinks/[SINK_ID]"
+     *
+     * Example: `"projects/my-project-id/sinks/my-sink-id"`.
      * @param array $optionalArgs {
      *                            Optional.
      *
@@ -635,19 +699,19 @@ class MetricsServiceV2GapicClient
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function deleteLogMetric($metricName, $optionalArgs = [])
+    public function deleteSink($sinkName, $optionalArgs = [])
     {
-        $request = new DeleteLogMetricRequest();
-        $request->setMetricName($metricName);
+        $request = new DeleteSinkRequest();
+        $request->setSinkName($sinkName);
 
-        $mergedSettings = $this->defaultCallSettings['deleteLogMetric']->merge(
+        $mergedSettings = $this->defaultCallSettings['deleteSink']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->metricsServiceV2Stub,
-            'DeleteLogMetric',
+            $this->configServiceV2Stub,
+            'DeleteSink',
             $mergedSettings,
-            $this->descriptors['deleteLogMetric']
+            $this->descriptors['deleteSink']
         );
 
         return $callable(
@@ -664,7 +728,7 @@ class MetricsServiceV2GapicClient
      */
     public function close()
     {
-        $this->metricsServiceV2Stub->close();
+        $this->configServiceV2Stub->close();
     }
 
     private function createCredentialsCallback()
