+++ src/Logging/V2/Gapic/MetricsServiceV2GapicClient.php
@@ -18,7 +18,7 @@
 /*
  * GENERATED CODE WARNING
  * This file was generated from the file
- * https://github.com/google/googleapis/blob/master/google/logging/v2/logging_config.proto
+ * https://github.com/google/googleapis/blob/master/google/logging/v2/logging_metrics.proto
  * and updates to that file get reflected here through a refresh process.
  *
  * EXPERIMENTAL: this client library class has not yet been declared beta. This class may change
@@ -28,7 +28,7 @@
  * @experimental
  */
 
-namespace Google\Cloud\Logging\V2;
+namespace Google\Cloud\Logging\V2\Gapic;
 
 use Google\GAX\AgentHeaderDescriptor;
 use Google\GAX\ApiCallable;
@@ -37,17 +37,16 @@ use Google\GAX\GrpcConstants;
 use Google\GAX\GrpcCredentialsHelper;
 use Google\GAX\PageStreamingDescriptor;
 use Google\GAX\PathTemplate;
-use Google\Logging\V2\ConfigServiceV2GrpcClient;
-use Google\Logging\V2\CreateSinkRequest;
-use Google\Logging\V2\DeleteSinkRequest;
-use Google\Logging\V2\GetSinkRequest;
-use Google\Logging\V2\ListSinksRequest;
-use Google\Logging\V2\LogSink;
-use Google\Logging\V2\UpdateSinkRequest;
+use Google\Logging\V2\CreateLogMetricRequest;
+use Google\Logging\V2\DeleteLogMetricRequest;
+use Google\Logging\V2\GetLogMetricRequest;
+use Google\Logging\V2\ListLogMetricsRequest;
+use Google\Logging\V2\LogMetric;
+use Google\Logging\V2\MetricsServiceV2GrpcClient;
+use Google\Logging\V2\UpdateLogMetricRequest;
 
 /**
- * Service Description: Service for configuring sinks used to export log entries outside of
- * Stackdriver Logging.
+ * Service Description: Service for configuring logs-based metrics.
  *
  * EXPERIMENTAL: this client library class has not yet been declared beta. This class may change
  * more frequently than those which have been declared beta or 1.0, including changes which break
@@ -58,23 +57,23 @@ use Google\Logging\V2\UpdateSinkRequest;
  *
  * ```
  * try {
- *     $configServiceV2Client = new ConfigServiceV2Client();
- *     $formattedParent = ConfigServiceV2Client::formatProjectName("[PROJECT]");
+ *     $metricsServiceV2Client = new MetricsServiceV2Client();
+ *     $formattedParent = MetricsServiceV2Client::formatProjectName("[PROJECT]");
  *     // Iterate through all elements
- *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent);
+ *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent);
  *     foreach ($pagedResponse->iterateAllElements() as $element) {
  *         // doSomethingWith($element);
  *     }
  *
  *     // OR iterate over pages of elements, with the maximum page size set to 5
- *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent, ['pageSize' => 5]);
+ *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent, ['pageSize' => 5]);
  *     foreach ($pagedResponse->iteratePages() as $page) {
  *         foreach ($page as $element) {
  *             // doSomethingWith($element);
  *         }
  *     }
  * } finally {
- *     $configServiceV2Client->close();
+ *     $metricsServiceV2Client->close();
  * }
  * ```
  *
@@ -85,7 +84,7 @@ use Google\Logging\V2\UpdateSinkRequest;
  *
  * @experimental
  */
-class ConfigServiceV2GapicClient
+class MetricsServiceV2GapicClient
 {
     /**
      * The default address of the service.
@@ -113,10 +112,10 @@ class ConfigServiceV2GapicClient
     const CODEGEN_VERSION = '0.0.5';
 
     private static $projectNameTemplate;
-    private static $sinkNameTemplate;
+    private static $metricNameTemplate;
 
     protected $grpcCredentialsHelper;
-    protected $configServiceV2Stub;
+    protected $metricsServiceV2Stub;
     private $scopes;
     private $defaultCallSettings;
     private $descriptors;
@@ -139,19 +138,19 @@ class ConfigServiceV2GapicClient
 
     /**
      * Formats a string containing the fully-qualified path to represent
-     * a sink resource.
+     * a metric resource.
      *
      * @param string $project
-     * @param string $sink
+     * @param string $metric
      *
-     * @return string The formatted sink resource.
+     * @return string The formatted metric resource.
      * @experimental
      */
-    public static function formatSinkName($project, $sink)
+    public static function formatMetricName($project, $metric)
     {
-        return self::getSinkNameTemplate()->render([
+        return self::getMetricNameTemplate()->render([
             'project' => $project,
-            'sink' => $sink,
+            'metric' => $metric,
         ]);
     }
 
@@ -171,30 +170,30 @@ class ConfigServiceV2GapicClient
 
     /**
      * Parses the project from the given fully-qualified path which
-     * represents a sink resource.
+     * represents a metric resource.
      *
-     * @param string $sinkName The fully-qualified sink resource.
+     * @param string $metricName The fully-qualified metric resource.
      *
      * @return string The extracted project value.
      * @experimental
      */
-    public static function parseProjectFromSinkName($sinkName)
+    public static function parseProjectFromMetricName($metricName)
     {
-        return self::getSinkNameTemplate()->match($sinkName)['project'];
+        return self::getMetricNameTemplate()->match($metricName)['project'];
     }
 
     /**
-     * Parses the sink from the given fully-qualified path which
-     * represents a sink resource.
+     * Parses the metric from the given fully-qualified path which
+     * represents a metric resource.
      *
-     * @param string $sinkName The fully-qualified sink resource.
+     * @param string $metricName The fully-qualified metric resource.
      *
-     * @return string The extracted sink value.
+     * @return string The extracted metric value.
      * @experimental
      */
-    public static function parseSinkFromSinkName($sinkName)
+    public static function parseMetricFromMetricName($metricName)
     {
-        return self::getSinkNameTemplate()->match($sinkName)['sink'];
+        return self::getMetricNameTemplate()->match($metricName)['metric'];
     }
 
     private static function getProjectNameTemplate()
@@ -206,29 +205,29 @@ class ConfigServiceV2GapicClient
         return self::$projectNameTemplate;
     }
 
-    private static function getSinkNameTemplate()
+    private static function getMetricNameTemplate()
     {
-        if (self::$sinkNameTemplate == null) {
-            self::$sinkNameTemplate = new PathTemplate('projects/{project}/sinks/{sink}');
+        if (self::$metricNameTemplate == null) {
+            self::$metricNameTemplate = new PathTemplate('projects/{project}/metrics/{metric}');
         }
 
-        return self::$sinkNameTemplate;
+        return self::$metricNameTemplate;
     }
 
     private static function getPageStreamingDescriptors()
     {
-        $listSinksPageStreamingDescriptor =
+        $listLogMetricsPageStreamingDescriptor =
                 new PageStreamingDescriptor([
                     'requestPageTokenGetMethod' => 'getPageToken',
                     'requestPageTokenSetMethod' => 'setPageToken',
                     'requestPageSizeGetMethod' => 'getPageSize',
                     'requestPageSizeSetMethod' => 'setPageSize',
                     'responsePageTokenGetMethod' => 'getNextPageToken',
-                    'resourcesGetMethod' => 'getSinks',
+                    'resourcesGetMethod' => 'getMetrics',
                 ]);
 
         $pageStreamingDescriptors = [
-            'listSinks' => $listSinksPageStreamingDescriptor,
+            'listLogMetrics' => $listLogMetricsPageStreamingDescriptor,
         ];
 
         return $pageStreamingDescriptors;
@@ -310,22 +309,22 @@ class ConfigServiceV2GapicClient
 
         $defaultDescriptors = ['headerDescriptor' => $headerDescriptor];
         $this->descriptors = [
-            'listSinks' => $defaultDescriptors,
-            'getSink' => $defaultDescriptors,
-            'createSink' => $defaultDescriptors,
-            'updateSink' => $defaultDescriptors,
-            'deleteSink' => $defaultDescriptors,
+            'listLogMetrics' => $defaultDescriptors,
+            'getLogMetric' => $defaultDescriptors,
+            'createLogMetric' => $defaultDescriptors,
+            'updateLogMetric' => $defaultDescriptors,
+            'deleteLogMetric' => $defaultDescriptors,
         ];
         $pageStreamingDescriptors = self::getPageStreamingDescriptors();
         foreach ($pageStreamingDescriptors as $method => $pageStreamingDescriptor) {
             $this->descriptors[$method]['pageStreamingDescriptor'] = $pageStreamingDescriptor;
         }
 
-        $clientConfigJsonString = file_get_contents(__DIR__.'/resources/config_service_v2_client_config.json');
+        $clientConfigJsonString = file_get_contents(__DIR__.'/../resources/metrics_service_v2_client_config.json');
         $clientConfig = json_decode($clientConfigJsonString, true);
         $this->defaultCallSettings =
                 CallSettings::load(
-                    'google.logging.v2.ConfigServiceV2',
+                    'google.logging.v2.MetricsServiceV2',
                     $clientConfig,
                     $options['retryingOverride'],
                     GrpcConstants::getStatusCodeNames(),
@@ -340,47 +339,44 @@ class ConfigServiceV2GapicClient
         }
         $this->grpcCredentialsHelper = new GrpcCredentialsHelper($options);
 
-        $createConfigServiceV2StubFunction = function ($hostname, $opts, $channel) {
-            return new ConfigServiceV2GrpcClient($hostname, $opts, $channel);
+        $createMetricsServiceV2StubFunction = function ($hostname, $opts, $channel) {
+            return new MetricsServiceV2GrpcClient($hostname, $opts, $channel);
         };
-        if (array_key_exists('createConfigServiceV2StubFunction', $options)) {
-            $createConfigServiceV2StubFunction = $options['createConfigServiceV2StubFunction'];
+        if (array_key_exists('createMetricsServiceV2StubFunction', $options)) {
+            $createMetricsServiceV2StubFunction = $options['createMetricsServiceV2StubFunction'];
         }
-        $this->configServiceV2Stub = $this->grpcCredentialsHelper->createStub($createConfigServiceV2StubFunction);
+        $this->metricsServiceV2Stub = $this->grpcCredentialsHelper->createStub($createMetricsServiceV2StubFunction);
     }
 
     /**
-     * Lists sinks.
+     * Lists logs-based metrics.
      *
      * Sample code:
      * ```
      * try {
-     *     $configServiceV2Client = new ConfigServiceV2Client();
-     *     $formattedParent = ConfigServiceV2Client::formatProjectName("[PROJECT]");
+     *     $metricsServiceV2Client = new MetricsServiceV2Client();
+     *     $formattedParent = MetricsServiceV2Client::formatProjectName("[PROJECT]");
      *     // Iterate through all elements
-     *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent);
+     *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent);
      *     foreach ($pagedResponse->iterateAllElements() as $element) {
      *         // doSomethingWith($element);
      *     }
      *
      *     // OR iterate over pages of elements, with the maximum page size set to 5
-     *     $pagedResponse = $configServiceV2Client->listSinks($formattedParent, ['pageSize' => 5]);
+     *     $pagedResponse = $metricsServiceV2Client->listLogMetrics($formattedParent, ['pageSize' => 5]);
      *     foreach ($pagedResponse->iteratePages() as $page) {
      *         foreach ($page as $element) {
      *             // doSomethingWith($element);
      *         }
      *     }
      * } finally {
-     *     $configServiceV2Client->close();
+     *     $metricsServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $parent Required. The parent resource whose sinks are to be listed:
+     * @param string $parent Required. The name of the project containing the metrics:
      *
      *     "projects/[PROJECT_ID]"
-     *     "organizations/[ORGANIZATION_ID]"
-     *     "billingAccounts/[BILLING_ACCOUNT_ID]"
-     *     "folders/[FOLDER_ID]"
      * @param array $optionalArgs {
      *                            Optional.
      *
@@ -406,9 +402,9 @@ class ConfigServiceV2GapicClient
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function listSinks($parent, $optionalArgs = [])
+    public function listLogMetrics($parent, $optionalArgs = [])
     {
-        $request = new ListSinksRequest();
+        $request = new ListLogMetricsRequest();
         $request->setParent($parent);
         if (isset($optionalArgs['pageToken'])) {
             $request->setPageToken($optionalArgs['pageToken']);
@@ -417,14 +413,14 @@ class ConfigServiceV2GapicClient
             $request->setPageSize($optionalArgs['pageSize']);
         }
 
-        $mergedSettings = $this->defaultCallSettings['listSinks']->merge(
+        $mergedSettings = $this->defaultCallSettings['listLogMetrics']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->configServiceV2Stub,
-            'ListSinks',
+            $this->metricsServiceV2Stub,
+            'ListLogMetrics',
             $mergedSettings,
-            $this->descriptors['listSinks']
+            $this->descriptors['listLogMetrics']
         );
 
         return $callable(
@@ -434,27 +430,22 @@ class ConfigServiceV2GapicClient
     }
 
     /**
-     * Gets a sink.
+     * Gets a logs-based metric.
      *
      * Sample code:
      * ```
      * try {
-     *     $configServiceV2Client = new ConfigServiceV2Client();
-     *     $formattedSinkName = ConfigServiceV2Client::formatSinkName("[PROJECT]", "[SINK]");
-     *     $response = $configServiceV2Client->getSink($formattedSinkName);
+     *     $metricsServiceV2Client = new MetricsServiceV2Client();
+     *     $formattedMetricName = MetricsServiceV2Client::formatMetricName("[PROJECT]", "[METRIC]");
+     *     $response = $metricsServiceV2Client->getLogMetric($formattedMetricName);
      * } finally {
-     *     $configServiceV2Client->close();
+     *     $metricsServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $sinkName Required. The resource name of the sink:
+     * @param string $metricName The resource name of the desired metric:
      *
-     *     "projects/[PROJECT_ID]/sinks/[SINK_ID]"
-     *     "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]"
-     *     "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]"
-     *     "folders/[FOLDER_ID]/sinks/[SINK_ID]"
-     *
-     * Example: `"projects/my-project-id/sinks/my-sink-id"`.
+     *     "projects/[PROJECT_ID]/metrics/[METRIC_ID]"
      * @param array $optionalArgs {
      *                            Optional.
      *
@@ -466,24 +457,24 @@ class ConfigServiceV2GapicClient
      *          is not set.
      * }
      *
-     * @return \Google\Logging\V2\LogSink
+     * @return \Google\Logging\V2\LogMetric
      *
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function getSink($sinkName, $optionalArgs = [])
+    public function getLogMetric($metricName, $optionalArgs = [])
     {
-        $request = new GetSinkRequest();
-        $request->setSinkName($sinkName);
+        $request = new GetLogMetricRequest();
+        $request->setMetricName($metricName);
 
-        $mergedSettings = $this->defaultCallSettings['getSink']->merge(
+        $mergedSettings = $this->defaultCallSettings['getLogMetric']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->configServiceV2Stub,
-            'GetSink',
+            $this->metricsServiceV2Stub,
+            'GetLogMetric',
             $mergedSettings,
-            $this->descriptors['getSink']
+            $this->descriptors['getLogMetric']
         );
 
         return $callable(
@@ -493,49 +484,30 @@ class ConfigServiceV2GapicClient
     }
 
     /**
-     * Creates a sink that exports specified log entries to a destination.  The
-     * export of newly-ingested log entries begins immediately, unless the current
-     * time is outside the sink's start and end times or the sink's
-     * `writer_identity` is not permitted to write to the destination.  A sink can
-     * export log entries only from the resource owning the sink.
+     * Creates a logs-based metric.
      *
      * Sample code:
      * ```
      * try {
-     *     $configServiceV2Client = new ConfigServiceV2Client();
-     *     $formattedParent = ConfigServiceV2Client::formatProjectName("[PROJECT]");
-     *     $sink = new LogSink();
-     *     $response = $configServiceV2Client->createSink($formattedParent, $sink);
+     *     $metricsServiceV2Client = new MetricsServiceV2Client();
+     *     $formattedParent = MetricsServiceV2Client::formatProjectName("[PROJECT]");
+     *     $metric = new LogMetric();
+     *     $response = $metricsServiceV2Client->createLogMetric($formattedParent, $metric);
      * } finally {
-     *     $configServiceV2Client->close();
+     *     $metricsServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $parent Required. The resource in which to create the sink:
+     * @param string $parent The resource name of the project in which to create the metric:
      *
      *     "projects/[PROJECT_ID]"
-     *     "organizations/[ORGANIZATION_ID]"
-     *     "billingAccounts/[BILLING_ACCOUNT_ID]"
-     *     "folders/[FOLDER_ID]"
      *
-     * Examples: `"projects/my-logging-project"`, `"organizations/123456789"`.
-     * @param LogSink $sink         Required. The new sink, whose `name` parameter is a sink identifier that
-     *                              is not already in use.
-     * @param array   $optionalArgs {
-     *                              Optional.
+     * The new metric must be provided in the request.
+     * @param LogMetric $metric       The new logs-based metric, which must not have an identifier that
+     *                                already exists.
+     * @param array     $optionalArgs {
+     *                                Optional.
      *
-     *     @type bool $uniqueWriterIdentity
-     *          Optional. Determines the kind of IAM identity returned as `writer_identity`
-     *          in the new sink.  If this value is omitted or set to false, and if the
-     *          sink's parent is a project, then the value returned as `writer_identity` is
-     *          the same group or service account used by Stackdriver Logging before the
-     *          addition of writer identities to this API. The sink's destination must be
-     *          in the same project as the sink itself.
-     *
-     *          If this field is set to true, or if the sink is owned by a non-project
-     *          resource such as an organization, then the value of `writer_identity` will
-     *          be a unique service account used only for exports from the new sink.  For
-     *          more information, see `writer_identity` in [LogSink][google.logging.v2.LogSink].
      *     @type \Google\GAX\RetrySettings $retrySettings
      *          Retry settings to use for this call. If present, then
      *          $timeoutMillis is ignored.
@@ -544,28 +516,25 @@ class ConfigServiceV2GapicClient
      *          is not set.
      * }
      *
-     * @return \Google\Logging\V2\LogSink
+     * @return \Google\Logging\V2\LogMetric
      *
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function createSink($parent, $sink, $optionalArgs = [])
+    public function createLogMetric($parent, $metric, $optionalArgs = [])
     {
-        $request = new CreateSinkRequest();
+        $request = new CreateLogMetricRequest();
         $request->setParent($parent);
-        $request->setSink($sink);
-        if (isset($optionalArgs['uniqueWriterIdentity'])) {
-            $request->setUniqueWriterIdentity($optionalArgs['uniqueWriterIdentity']);
-        }
+        $request->setMetric($metric);
 
-        $mergedSettings = $this->defaultCallSettings['createSink']->merge(
+        $mergedSettings = $this->defaultCallSettings['createLogMetric']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->configServiceV2Stub,
-            'CreateSink',
+            $this->metricsServiceV2Stub,
+            'CreateLogMetric',
             $mergedSettings,
-            $this->descriptors['createSink']
+            $this->descriptors['createLogMetric']
         );
 
         return $callable(
@@ -575,54 +544,31 @@ class ConfigServiceV2GapicClient
     }
 
     /**
-     * Updates a sink. If the named sink doesn't exist, then this method is
-     * identical to
-     * [sinks.create](https://cloud.google.com/logging/docs/api/reference/rest/v2/projects.sinks/create).
-     * If the named sink does exist, then this method replaces the following
-     * fields in the existing sink with values from the new sink: `destination`,
-     * `filter`, `output_version_format`, `start_time`, and `end_time`.
-     * The updated filter might also have a new `writer_identity`; see the
-     * `unique_writer_identity` field.
+     * Creates or updates a logs-based metric.
      *
      * Sample code:
      * ```
      * try {
-     *     $configServiceV2Client = new ConfigServiceV2Client();
-     *     $formattedSinkName = ConfigServiceV2Client::formatSinkName("[PROJECT]", "[SINK]");
-     *     $sink = new LogSink();
-     *     $response = $configServiceV2Client->updateSink($formattedSinkName, $sink);
+     *     $metricsServiceV2Client = new MetricsServiceV2Client();
+     *     $formattedMetricName = MetricsServiceV2Client::formatMetricName("[PROJECT]", "[METRIC]");
+     *     $metric = new LogMetric();
+     *     $response = $metricsServiceV2Client->updateLogMetric($formattedMetricName, $metric);
      * } finally {
-     *     $configServiceV2Client->close();
+     *     $metricsServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $sinkName Required. The full resource name of the sink to update, including the
-     *                         parent resource and the sink identifier:
+     * @param string $metricName The resource name of the metric to update:
      *
-     *     "projects/[PROJECT_ID]/sinks/[SINK_ID]"
-     *     "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]"
-     *     "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]"
-     *     "folders/[FOLDER_ID]/sinks/[SINK_ID]"
+     *     "projects/[PROJECT_ID]/metrics/[METRIC_ID]"
      *
-     * Example: `"projects/my-project-id/sinks/my-sink-id"`.
-     * @param LogSink $sink         Required. The updated sink, whose name is the same identifier that appears
-     *                              as part of `sink_name`.  If `sink_name` does not exist, then
-     *                              this method creates a new sink.
-     * @param array   $optionalArgs {
-     *                              Optional.
+     * The updated metric must be provided in the request and it's
+     * `name` field must be the same as `[METRIC_ID]` If the metric
+     * does not exist in `[PROJECT_ID]`, then a new metric is created.
+     * @param LogMetric $metric       The updated metric.
+     * @param array     $optionalArgs {
+     *                                Optional.
      *
-     *     @type bool $uniqueWriterIdentity
-     *          Optional. See
-     *          [sinks.create](https://cloud.google.com/logging/docs/api/reference/rest/v2/projects.sinks/create)
-     *          for a description of this field.  When updating a sink, the effect of this
-     *          field on the value of `writer_identity` in the updated sink depends on both
-     *          the old and new values of this field:
-     *
-     *          +   If the old and new values of this field are both false or both true,
-     *              then there is no change to the sink's `writer_identity`.
-     *          +   If the old value is false and the new value is true, then
-     *              `writer_identity` is changed to a unique service account.
-     *          +   It is an error if the old value is true and the new value is false.
      *     @type \Google\GAX\RetrySettings $retrySettings
      *          Retry settings to use for this call. If present, then
      *          $timeoutMillis is ignored.
@@ -631,28 +577,25 @@ class ConfigServiceV2GapicClient
      *          is not set.
      * }
      *
-     * @return \Google\Logging\V2\LogSink
+     * @return \Google\Logging\V2\LogMetric
      *
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function updateSink($sinkName, $sink, $optionalArgs = [])
+    public function updateLogMetric($metricName, $metric, $optionalArgs = [])
     {
-        $request = new UpdateSinkRequest();
-        $request->setSinkName($sinkName);
-        $request->setSink($sink);
-        if (isset($optionalArgs['uniqueWriterIdentity'])) {
-            $request->setUniqueWriterIdentity($optionalArgs['uniqueWriterIdentity']);
-        }
+        $request = new UpdateLogMetricRequest();
+        $request->setMetricName($metricName);
+        $request->setMetric($metric);
 
-        $mergedSettings = $this->defaultCallSettings['updateSink']->merge(
+        $mergedSettings = $this->defaultCallSettings['updateLogMetric']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->configServiceV2Stub,
-            'UpdateSink',
+            $this->metricsServiceV2Stub,
+            'UpdateLogMetric',
             $mergedSettings,
-            $this->descriptors['updateSink']
+            $this->descriptors['updateLogMetric']
         );
 
         return $callable(
@@ -662,29 +605,22 @@ class ConfigServiceV2GapicClient
     }
 
     /**
-     * Deletes a sink. If the sink has a unique `writer_identity`, then that
-     * service account is also deleted.
+     * Deletes a logs-based metric.
      *
      * Sample code:
      * ```
      * try {
-     *     $configServiceV2Client = new ConfigServiceV2Client();
-     *     $formattedSinkName = ConfigServiceV2Client::formatSinkName("[PROJECT]", "[SINK]");
-     *     $configServiceV2Client->deleteSink($formattedSinkName);
+     *     $metricsServiceV2Client = new MetricsServiceV2Client();
+     *     $formattedMetricName = MetricsServiceV2Client::formatMetricName("[PROJECT]", "[METRIC]");
+     *     $metricsServiceV2Client->deleteLogMetric($formattedMetricName);
      * } finally {
-     *     $configServiceV2Client->close();
+     *     $metricsServiceV2Client->close();
      * }
      * ```
      *
-     * @param string $sinkName Required. The full resource name of the sink to delete, including the
-     *                         parent resource and the sink identifier:
+     * @param string $metricName The resource name of the metric to delete:
      *
-     *     "projects/[PROJECT_ID]/sinks/[SINK_ID]"
-     *     "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]"
-     *     "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]"
-     *     "folders/[FOLDER_ID]/sinks/[SINK_ID]"
-     *
-     * Example: `"projects/my-project-id/sinks/my-sink-id"`.
+     *     "projects/[PROJECT_ID]/metrics/[METRIC_ID]"
      * @param array $optionalArgs {
      *                            Optional.
      *
@@ -699,19 +635,19 @@ class ConfigServiceV2GapicClient
      * @throws \Google\GAX\ApiException if the remote call fails
      * @experimental
      */
-    public function deleteSink($sinkName, $optionalArgs = [])
+    public function deleteLogMetric($metricName, $optionalArgs = [])
     {
-        $request = new DeleteSinkRequest();
-        $request->setSinkName($sinkName);
+        $request = new DeleteLogMetricRequest();
+        $request->setMetricName($metricName);
 
-        $mergedSettings = $this->defaultCallSettings['deleteSink']->merge(
+        $mergedSettings = $this->defaultCallSettings['deleteLogMetric']->merge(
             new CallSettings($optionalArgs)
         );
         $callable = ApiCallable::createApiCall(
-            $this->configServiceV2Stub,
-            'DeleteSink',
+            $this->metricsServiceV2Stub,
+            'DeleteLogMetric',
             $mergedSettings,
-            $this->descriptors['deleteSink']
+            $this->descriptors['deleteLogMetric']
         );
 
         return $callable(
@@ -728,7 +664,7 @@ class ConfigServiceV2GapicClient
      */
     public function close()
     {
-        $this->configServiceV2Stub->close();
+        $this->metricsServiceV2Stub->close();
     }
 
     private function createCredentialsCallback()
