+++ src/Monitoring/V3/GroupServiceClient.php
@@ -30,6 +30,8 @@
 
 namespace Google\Cloud\Monitoring\V3;
 
+use Google\Cloud\Monitoring\V3\Gapic\GroupServiceGapicClient;
+
 /**
  * {@inheritdoc}
  */
