+++ src/Spanner/Admin/Database/V1/Gapic/DatabaseAdminGapicClient.php
@@ -28,7 +28,7 @@
  * @experimental
  */
 
-namespace Google\Cloud\Spanner\Admin\Database\V1;
+namespace Google\Cloud\Spanner\Admin\Database\V1\Gapic;
 
 use Google\GAX\AgentHeaderDescriptor;
 use Google\GAX\ApiCallable;
@@ -432,7 +432,7 @@ class DatabaseAdminGapicClient
             $this->descriptors[$method]['longRunningDescriptor'] = $longRunningDescriptor + ['operationsClient' => $this->operationsClient];
         }
 
-        $clientConfigJsonString = file_get_contents(__DIR__.'/resources/database_admin_client_config.json');
+        $clientConfigJsonString = file_get_contents(__DIR__.'/../resources/database_admin_client_config.json');
         $clientConfig = json_decode($clientConfigJsonString, true);
         $this->defaultCallSettings =
                 CallSettings::load(
