+++ src/Spanner/V1/resources/spanner_client_config.json
@@ -6,7 +6,10 @@
           "DEADLINE_EXCEEDED",
           "UNAVAILABLE"
         ],
-        "non_idempotent": []
+        "non_idempotent": [],
+        "long_running": [
+          "UNAVAILABLE"
+        ]
       },
       "retry_params": {
         "default": {
@@ -17,12 +20,21 @@
           "rpc_timeout_multiplier": 1.0,
           "max_rpc_timeout_millis": 60000,
           "total_timeout_millis": 600000
+        },
+        "long_running": {
+          "initial_retry_delay_millis": 1000,
+          "retry_delay_multiplier": 1.3,
+          "max_retry_delay_millis": 32000,
+          "initial_rpc_timeout_millis": 3600000,
+          "rpc_timeout_multiplier": 1.0,
+          "max_rpc_timeout_millis": 3600000,
+          "total_timeout_millis": 36000000
         }
       },
       "methods": {
         "CreateSession": {
           "timeout_millis": 30000,
-          "retry_codes_name": "non_idempotent",
+          "retry_codes_name": "idempotent",
           "retry_params_name": "default"
         },
         "GetSession": {
@@ -37,33 +49,33 @@
         },
         "ExecuteSql": {
           "timeout_millis": 30000,
-          "retry_codes_name": "non_idempotent",
+          "retry_codes_name": "idempotent",
           "retry_params_name": "default"
         },
         "ExecuteStreamingSql": {
-          "timeout_millis": 30000
+          "timeout_millis": 3600000
         },
         "Read": {
           "timeout_millis": 30000,
-          "retry_codes_name": "non_idempotent",
+          "retry_codes_name": "idempotent",
           "retry_params_name": "default"
         },
         "StreamingRead": {
-          "timeout_millis": 30000
+          "timeout_millis": 3600000
         },
         "BeginTransaction": {
           "timeout_millis": 30000,
-          "retry_codes_name": "non_idempotent",
+          "retry_codes_name": "idempotent",
           "retry_params_name": "default"
         },
         "Commit": {
-          "timeout_millis": 30000,
-          "retry_codes_name": "non_idempotent",
-          "retry_params_name": "default"
+          "timeout_millis": 3600000,
+          "retry_codes_name": "long_running",
+          "retry_params_name": "long_running"
         },
         "Rollback": {
           "timeout_millis": 30000,
-          "retry_codes_name": "non_idempotent",
+          "retry_codes_name": "idempotent",
           "retry_params_name": "default"
         }
       }
